+++ public/assets/scripts/game.js
@@ -1,6 +1,6 @@
 import { Property } from "./models/property.js";
 import { Map } from "./models/map.js";
-import { Dice } from "./models/dice.js";
+import { Die } from "./models/die.js";
 
 const boardBox = document.getElementById("board-wrapper");
 var canvas = document.getElementById("game-board");
@@ -13,9 +13,9 @@ map.setCanvasSize();
 var properties = Property.getInitialProperties();
 map.drawProperty();
 
-var dice = Dice.rollSingleDice();
-var dualDice = Dice.rollDice();
-console.log(dice.value);
-console.log(dualDice);
+var die = Die.rollDie();
+var dice = Die.rollDice();
+console.log(die.value);
+console.log(dice);
 console.log(properties);
 console.log(map);
